+++ src/types/typings.ts
@@ -135,6 +135,7 @@ export type UserEvent =
   | 'forgot-username-attempt'
   | 'email-change-attempt'
   | 'user-db-added'
+  | 'user-deleting'
   | 'user-deleted'
   | 'confirmation-email-error';