+++ src/user-hashing.ts
@@ -1,75 +1,80 @@
 import pwdModule from '@sl-nx/couch-pwd';
 import { Config } from './types/config';
-import { HashResult, LocalHashObj } from './types/typings';
-import { URLSafeUUID } from './util';
-
-const pwd = new pwdModule();
+import { HashResult } from './types/typings';
+import { UserHashingLegacy } from './user-hashing-legacy';
 
 /**
  * Class for hashing and verifying sl-user passwords
  */
 export class UserHashing {
-  hashers = [];
-  times: number[] = [];
-  dummyHashObject: LocalHashObj = { iterations: 10 };
+  private legacy: UserHashingLegacy;
+
+  private pwdCouch: pwdModule;
+  private iterations: number;
+  private pbkdf2Prf: string;
+  private keyLength: number;
+  private saltLength: number;
 
   constructor(config: Partial<Config>) {
-    const iterationPairs = config.security?.iterations;
-    if (iterationPairs) {
-      for (const pair of config.security.iterations) {
-        this.times.push(pair[0]);
-        this.hashers.push(new pwdModule(pair[1]));
-      }
-    }
-    this.hashUserPassword(URLSafeUUID()).then(dummy => {
-      this.dummyHashObject = dummy;
-    });
+    this.legacy = new UserHashingLegacy(config);
+    this.iterations = config.security?.userHashing?.iterations || 600000;
+    this.pbkdf2Prf = config.security?.userHashing?.pbkdf2Prf || 'sha256';
+    this.keyLength = config.security?.userHashing?.keyLength || (this.pbkdf2Prf === 'sha' ? 20 : 32);
+    this.saltLength = config.security?.userHashing?.saltLength || 16;
+
+    this.pwdCouch = UserHashing.createPwdModule(
+      this.iterations,
+      this.keyLength,
+      this.saltLength,
+      this.pbkdf2Prf
+    );
   }
 
-  private getHasherForTimestamp(ts: number = undefined) {
-    let ret = pwd;
-    if (this.times.length === 0 || ts === undefined) {
-      return ret;
+  isUpgradeNeeded(hashObj: HashResult): boolean {
+    if (hashObj.iterations === undefined) {
+      return true;
     }
-    for (let idx = 0; idx < this.times.length; idx++) {
-      if (ts > this.times[idx]) {
-        ret = this.hashers[idx];
-      } else {
-        break;
-      }
+    if (hashObj.iterations < this.iterations) {
+      return true;
     }
-    return ret;
+    if ((hashObj.pbkdf2_prf || 'sha') !== this.pbkdf2Prf) {
+      return true;
+    }
+    return false;
   }
 
-  hashUserPassword(pw: string): Promise<HashResult> {
-    const t = new Date().valueOf();
+  hashUserPassword(password: string): Promise<HashResult> {
     return new Promise((resolve, reject) => {
-      this.getHasherForTimestamp(t).hash(pw, (err, salt, hash) => {
+      this.pwdCouch.hash(password, (err, salt, hash) => {
         if (err) {
           return reject(err);
         }
         return resolve({
+          created: Date.now(),
           salt: salt,
-          derived_key: hash
+          derived_key: hash,
+          password_scheme: 'pbkdf2',
+          pbkdf2_prf: this.pbkdf2Prf,
+          iterations: this.iterations
         });
       });
-    }).then((hr: HashResult) => {
-      hr.created = t;
-      return hr;
     });
   }
 
   verifyUserPassword(hashObj: HashResult, pw: string): Promise<boolean> {
-    const salt = hashObj.salt ?? this.dummyHashObject.salt;
-    const derived_key = hashObj.derived_key ?? this.dummyHashObject.derived_key;
-    let created = hashObj.created;
-    if (!hashObj.salt && !hashObj.derived_key) {
-      created = this.dummyHashObject.created;
+    if (hashObj.iterations === undefined) {
+      return this.legacy.verifyUserPassword(hashObj, pw);
     }
 
     return new Promise((resolve, reject) => {
-      const hasher = this.getHasherForTimestamp(created);
-      hasher.hash(pw, salt, (err, hash) => {
+      const iterations = hashObj.iterations || 10;
+      const digest = hashObj.pbkdf2_prf || 'sha';
+      const length = digest === 'sha' ? 20 : 32;
+      const pwdCouch = UserHashing.createPwdModule(iterations, length, 16, digest);
+      
+      const salt = hashObj.salt;
+      const derived_key = hashObj.derived_key;
+      pwdCouch.hash(pw, salt, (err, hash) => {
         if (err) {
           return reject(err);
         } else if (hash !== derived_key) {
@@ -79,5 +84,15 @@ export class UserHashing {
         }
       });
     });
+  }
+
+  private static createPwdModule(iterations: number, keyLength: number, saltLength: number, digest: string): pwdModule {
+    return new pwdModule(
+      iterations,
+      keyLength,
+      saltLength,
+      'hex',
+      digest === 'sha' ? 'sha1' : digest
+    );
   }
 }